refactor: refactor model and buffer
